requirements.txt
fastai==2.7.12
keras==2.12.0
numpy==1.23.5
Pillow==9.4.0
Pillow==9.5.0
streamlit==1.22.0
tensorflow==2.12.0
tensorflow_intel==2.12.0
